Make `Consumable` consistent with other equipment
`Consumable` looks to be the only type of equipment with a category that
didn't have its category in it. To make it more consistent with the
other equipment, we put the category in `Consumable`.

This change lets the `Wrapper` type become simpler, since it doesn't
need to account for this one different case.

Also, since we don't have to account for results coming back in
different shapes, we don't have to worry about handling the different
cases in `flatMap`, and can use `Array.flat` instead.