Require `DocumentInitParameters` for `tokenizePDF`
The way to `pdf.js` works, it doesn't support CMaps by default. In order
to support more than the original Core Rulebook PDF, we need to enable
CMaps.

With `tokenizePDF` taking a bunch of different types as input, that
makes things a little bit harder. We'd have to case on each different
type to figure out what it is, partially build up a
`DocumentInitParameters`, and then default the CMap stuff; all so that
one of the two calls to `tokenizePDF` can pass in a `string`.

Rather than continue with that, we force that one call to pass in a
`DocumentInitParameters`. It's a little more confusing for that one
place, but should keep the complexity down a bit in `tokenizePDF`.