Support CMaps in `tokenizePDF`
Some of the other PDFs use non-Latin fonts. `pdf.js` doesn't support
these fonts by default. So, some of the content ends up missing. CMaps
enable support for non-Latin fonts, but they have to be enabled
explicitly in `pdf.js`.

In order to enable this, we have to point to a "URL" for where the CMaps
live. As it turns out, `pdfjs-dist` actually ships with the CMaps it
needs; which raises the question: why can't it do this by default, or at
least make it easier?

In any case, when running in Node.js (local validation or running the
tests) we use the CMaps we've already downloaded in the `node_modules`
directory. When running in the browser (foundryvtt) we use the CMaps
from the same version as what we're using. This should make it so we
don't have a mismatch as the version of `pdfjs-dist` changes over time.

While it would have been nice to use the stuff that's in the
`node_modules` directory no matter where the code is running, there
didn't seem to be a straight-forward way to make that work with the
current bundling setup. It's surely possible given enough time and
effort, but this alternative ought to work out okay instead.

With this in place, we should be able to start importing more PDFs.