Split `PAGES`
We want to move the definition of what type of pages things are to a
single location. But that location doesn't need to know things about how
to save something in FoundryVTT.

To separate out the definition of what each page contains, and how to
process that in different situations, we separate out `PAGES` into its
parts.

This separation also lets us start to talk about the "content" of a
page. This might lead to us being able to deal with pages with more than
one kind of content, like some of the pages of the atlases. No real
ideas there yet, but something to think about.