Remove `Wrapper`
We can break apart what `Wrapper` was doing and have the constituent
pieces of logic available to mix and match.

Even though this simplifies how `parsePdf` works–it no longer needs to
invert control in order to parse a page more generally–that's not really
the money of this change. The real benefit is that the association of
what content is on a page, and what parser will work for that page is
now able to be extracted out of `src/apps/import-pdf.ts` and someplace
more general.