Use a `Map<_, _>` instead of a `Record<_, _>`
Although writing `Record<_, _>` is convenient, iterating over it is not
quite that straight-forward. Since it's implemented as an object
literal, iterating over its entries with `Object.entries` gives all of
the keys as strings (since anything can be an objects own enumerable
property). So we have to do a dance where we attempt to parse a
`string`–that we know should be a `number`–back into a `number` in order
to use it.

Rather than deal with all that, we switch to using `Map<_, _>` so that
it can keep its keys the same type as they already are and not have to
deal with losing type information.