Use `test.each`
## Context

`jest` comes with a way to setup table tests. Rather than roll our own
thing, we use the built-in `test.each` function to setup the tests.

## Testing

We have to do a little munging of the data, but otherwise this should be
the same as before. So if the tests are passing, we should have
confidence that this didn't change any testing behavior in a material
way.