Use `describe.each`
## Context

We want to start testing multiple PDFs. So we switch the top `describe`
to being a table test. This should let us run through each PDF and test
them.

## Testing

Since this only changes how the `describe` is defined. So long as the
tests continue to pass, it should give us enough confidence that we
haven't changed the behavior in a material way.