Don't throw for non-existent PDFs
## Context

We want to start testing multiple PDFs, but not everyone will have all
of the PDFs that we might test. For those cases, we don't need to fail
if they're not all there. We can test as many that exist, and ignore any
PDFs that don't exist.

We change the initialization to check if we've got an `ENOENT` error
(which means the file doesn't exist). If it does, we return `null` and
skip setting up any actual tests.

## Testing

The interesting thing about `jest` is that it will require each test
suite to have at least one test. So if there are no PDfs at all, it wil
fail the test suite. But if there's at least one PDF, it will run and
ignore the ones that don't exist.

This behavior of ignoring a PDF that doesn't exist is effectively the
same as it was before. Since the `fs.readFileSync` would throw and stop
the tests before if the PDF didn't exist, and now `jest` will fail the
suite if there are no tests (because we exited early for lack of a PDF).