Test both Core Rulebook versions
## Context

With everything in place, we can test both the original Core Rulebook
and version 1.02 of the Core Rulebook.

Since the tests are now setup to have at least one `describe` with some
tests in them, this shouldn't require that both versions of the Core
Rulebook exist on the file system, only one of them.

## Testing

If the tests pass, this should give us enough confidence that this works
properly.