Name the PDFs
## Context

We want to start supporting more than one PDF. In order to do that, it's
a little easier if we can explicitly call out which PDF we're working
with.

We start by naming the only PDF that has explicit support: the Core
Rulebook. We'll add more PDFs as we extend support for them.

## Testing

This doesn't really change anything. So the tests continuing to pass is
a good enough signal that nothing has broken.