Allow selecting PDF type when importing
## Context

Since there doesn't really seem to be a definitive way to determine
which type of PDF we're supposed to be importing, we take the easy way
out and require the person importing the PDF to choose which type of PDF
it is.

Right now, it's defaulted to the Core Rulebook. And the only option is
to select the Core Rulebook. We'll add more types of PDFs and then be
able to import them (with their different pages and structure and
whatever else).

## Testing

The easiest way to validate that this works is to load this up in
Foundry, and import a PDF. This should all work as before.